requirements.txt
anyio==3.6.2
APScheduler==3.6.3
attrs==23.1.0
cachetools==4.2.2
certifi==2022.12.7
charset-normalizer==3.1.0
DateTime==5.5
DAWG-Python==0.7.2
docopt==0.6.2
et-xmlfile==1.1.0
geojson==2.5.0
h11==0.14.0
h2==4.1.0
hpack==4.0.0
httpcore==0.16.3
httpx==0.24.1
hyperframe==6.0.1
idna==3.4
isodate==0.6.1
lxml==4.9.3
openpyxl==3.1.2
platformdirs==3.8.1
pyaspeller==1.1.0
pymorphy2==0.9.1
pymorphy2-dicts-ru==2.4.417127.4579844
pyowm==3.3.0
PySocks==1.7.1
pyTelegramBotAPI==4.10.0
python-telegram-bot==13.7
pytz==2023.3
requests==2.28.2
requests-file==1.5.1
requests-toolbelt==1.0.0
rfc3986==1.5.0
schedule==1.2.0
six==1.16.0
sniffio==1.3.0
tornado==6.3.2
tzdata==2023.3
tzlocal==5.0.1
urllib3==1.26.15
zeep==4.2.1
zope.interface==6.2
